Implemented AuthCodeVerificationController version 2:
We now return two JWT tokens instead of one: One for the red (SwissCovid), one for the purple (NotifyMe) backend, which can then be used to upload GaenKeys or TraceKeys, respectively.
The tokens are returned in an AuthorizationCodeVerifyDtoWrapper, which is a simple Pojo containing a list of AuthorizationCodeVerifyDto objects.
Additional testcases are also included, although the ControllerSecurityTest is ommitted as it doesn't actually seem to do anything different to the normal ControllerTest.